refactor: use button to redirect to result page
Using the useRouter() hook for programmatic navigation seemed rather
unnatural. Instead, a button was added to the bottom of the active quiz
page to redirect to the results page.

Furthermore, the button is disabled until all the questions have been
attempted - this will prevent the user from prematurely entering the
results page.