refactor: move button styles to sass files
Doing this enables re-use of the button styles across two different
views: QuizView (the page in which options are chosen) and the
QuizGameView, where the quiz is actually played.

Two variations are used: enabled and disabled. Normal buttons use both
the base class and the enabled class, whereas the dynamic button in the
QuizGameView view uses both enabled and disabled classes.

A separate folder called 'components' is made inside the styles folder
for styling all the buttons over the site.