refactor: move all button styles to sass file
All the styles for 'transparent' button were moved to a separate
buttons.scss file to ensure separation of concerns.

By doing this, buttons in the navbar and in the hero section of the
home page were refactored. These buttons were also moved into a
separate div container to avoid some weird overflow issues.