refactor: styles in buttons scss file
The button in the quiz results view used a custom style defined in the
Vue SFC file itself: this has been replaced with the button defined in
the buttons.scss file.

Since this kind of button had a text transform of uppercase attached
to it, all the text of the buttons which made of use of this particular
class were manually re-written to use uppercase.