feat: make image heights fluid in cards
The circular images in preparations steps component was too big on
smaller screens, so their widths were made fluid using clamp.

Also, the large vertical image in the first row was provided with a
fluid height.

The largest screen size considered for this was 1440px, with the
smallest being 320px. The size of all the images is double on the
largest screen compared to that of the smallest.