feat: dark mode, some changes to PreparationSteps
This involved adding another colour variable for pink background in
light mode.

Also, one of the classes previously used in this component used the name
'dark' - this was renamed to 'contrast' to make a distinction in both
light and dark modes.

Also, the images were not not being centered properly in the step card.
This was fixed by adding an image container and centering the contents
of said container.