feat: show questions one at a time
A new view called the SingleQuestionView has been introduced, which is
responsible for this.

Whenever the user starts the quiz, the api is called. That is, the api
is called only when the value of the id param in the dynamic route is
equal to 1.

An temporary button was also added to navigate to the next question.

To fetch the corresponding question, a getter function was written in
the questions Pinia store. It fetches the (id - 1)th question in the
pinia store.

Also, to force re-renders of the dynamic route, a key was provided to
the router view.

Finally, because of the change in the page route in which the quiz
actually takes place, the button in the quiz options page now links
to the new route.

The old view which shows all the 20 questions at once has not been
removed, however. It is accessible by typing in the url but cannot
be navigated to by just clicking on the links in the app.