feat: add animations to hero and question cards
This was done by using the useMotion() hook provided by vueuse motion.
This hook was further extracted into more hooks for the preferred
animation.

Plus, to make the animation a bit better, the entire home view was
wrapped in the content grid class and provided an overflow value of
hidden.