feat: add vue router transitions, change timing fxn
This was done by using a dynamic component, and wrappping it within
a Transition tag. A key of the current path was provided to the current
component in order to both make the animations work and to force a
page refresh when switching between dynamic routes.

Also change the timing functions in the animations hooks.